requirements.txt
# Core framework
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# retriever
llama-index-core>=0.10.0
llama-index-graph-stores-neo4j
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-retrievers-bm25
faiss-cpu>=1.7.0
pypdf>=3.0.0
rank_bm25


# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client (vLLM remote calls)
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
nest-asyncio

# UI
gradio